TINY-6600: Testing resize prevention in BarManager and Bars
