+++ modules/snooker/src/main/ts/ephox/snooker/api/TableResize.ts
@@ -1,4 +1,3 @@
-import { Arr, Optional } from '@ephox/katamari';
 import { Bindable, Event, Events } from '@ephox/porkbun';
 import { SugarElement } from '@ephox/sugar';
 import * as Adjustments from '../resize/Adjustments';
@@ -7,7 +6,6 @@ import * as BarPositions from '../resize/BarPositions';
 import { ResizeBehaviour } from './ResizeBehaviour';
 import { ResizeWire } from './ResizeWire';
 import { TableSize } from './TableSize';
-import { Warehouse } from './Warehouse';
 
 type BarPositions<A> = BarPositions.BarPositions<A>;
 type ResizeType = 'row' | 'col';
@@ -49,23 +47,23 @@ export interface TableResize {
 // const isResizable = (elm: SugarElement<Element>) => Attribute.get(elm, 'data-mce-resize') !== 'false';
 
 // Checking cells could be an issue with a number column as that cell in the row will have the data-mce-resize set preventing any of the rows from being resized
-const canResizeRow = (table: SugarElement<HTMLTableElement>, rowIndex: number, canResize: (elm: SugarElement<Element>) => boolean) => {
-  const warehouse = Warehouse.fromTable(table);
-  const rowOpt = Optional.from(warehouse.all[rowIndex]);
-  const editableRow = rowOpt.map((row) => row.element).forall(canResize);
-  const editableCells = rowOpt.map((row) => row.cells).forall((cells) => Arr.forall(cells, (cell) => canResize(cell.element)));
-  return editableRow && editableCells;
-};
+// const canResizeRow = (table: SugarElement<HTMLTableElement>, rowIndex: number, canResize: (elm: SugarElement<Element>) => boolean) => {
+//   const warehouse = Warehouse.fromTable(table);
+//   const rowOpt = Optional.from(warehouse.all[rowIndex]);
+//   const editableRow = rowOpt.map((row) => row.element).forall(canResize);
+//   const editableCells = rowOpt.map((row) => row.cells).forall((cells) => Arr.forall(cells, (cell) => canResize(cell.element)));
+//   return editableRow && editableCells;
+// };
 
-// Should I also check colgroup element if it exists?
-const canResizeColumn = (table: SugarElement<HTMLTableElement>, columnIndex: number, canResize: (elm: SugarElement<Element>) => boolean) => {
-  const warehouse = Warehouse.fromTable(table);
-  // const editableCol = Warehouse.getColumnAt(warehouse, columnIndex).map((col) => col.element).forall(canResize);
-  const editableCol = Warehouse.getColumnAt(warehouse, columnIndex).map((col) => col.element).forall(canResize);
-  const columnCells = Warehouse.filterItems(warehouse, (cell) => cell.column === columnIndex);
-  const editableCells = Arr.forall(columnCells, (cell) => canResize(cell.element));
-  return editableCol && editableCells;
-};
+// // Should I also check colgroup element if it exists?
+// const canResizeColumn = (table: SugarElement<HTMLTableElement>, columnIndex: number, canResize: (elm: SugarElement<Element>) => boolean) => {
+//   const warehouse = Warehouse.fromTable(table);
+//   // const editableCol = Warehouse.getColumnAt(warehouse, columnIndex).map((col) => col.element).forall(canResize);
+//   const editableCol = Warehouse.getColumnAt(warehouse, columnIndex).map((col) => col.element).forall(canResize);
+//   const columnCells = Warehouse.filterItems(warehouse, (cell) => cell.column === columnIndex);
+//   const editableCells = Arr.forall(columnCells, (cell) => canResize(cell.element));
+//   return editableCol && editableCells;
+// };
 
 const create = (wire: ResizeWire, resizing: ResizeBehaviour, lazySizing: (element: SugarElement<HTMLTableElement>) => TableSize, canResize: (elm: SugarElement<Element>) => boolean): TableResize => {
   const hdirection = BarPositions.height;
@@ -80,13 +78,13 @@ const create = (wire: ResizeWire, resizing: ResizeBehaviour, lazySizing: (elemen
 
   manager.events.adjustHeight.bind((event) => {
     const table = event.table;
-    if (canResizeRow(table, event.row, canResize)) {
-      events.trigger.beforeResize(table, 'row');
-      const delta = hdirection.delta(event.delta, table);
-      // TODO: Use the resizing behaviour for heights as well
-      Adjustments.adjustHeight(table, delta, event.row, hdirection);
-      events.trigger.afterResize(table, 'row');
-    }
+    // if (canResizeRow(table, event.row, canResize)) {
+    // }
+    events.trigger.beforeResize(table, 'row');
+    const delta = hdirection.delta(event.delta, table);
+    // TODO: Use the resizing behaviour for heights as well
+    Adjustments.adjustHeight(table, delta, event.row, hdirection);
+    events.trigger.afterResize(table, 'row');
   });
 
   manager.events.startAdjust.bind((_event) => {
@@ -96,13 +94,13 @@ const create = (wire: ResizeWire, resizing: ResizeBehaviour, lazySizing: (elemen
   manager.events.adjustWidth.bind((event) => {
     const table = event.table;
     // Have a canResizeColumn function?
-    if (canResizeColumn(table, event.column, canResize)) {
-      events.trigger.beforeResize(table, 'col');
-      const delta = vdirection.delta(event.delta, table);
-      const tableSize = lazySizing(table);
-      Adjustments.adjustWidth(table, delta, event.column, resizing, tableSize);
-      events.trigger.afterResize(table, 'col');
-    }
+    // if (canResizeColumn(table, event.column, canResize)) {
+    // }
+    events.trigger.beforeResize(table, 'col');
+    const delta = vdirection.delta(event.delta, table);
+    const tableSize = lazySizing(table);
+    Adjustments.adjustWidth(table, delta, event.column, resizing, tableSize);
+    events.trigger.afterResize(table, 'col');
   });
 
   return {
